enable completions
- show completions after typing '.'
- show competions in demand when pressing 'ctrl+space'

Completions responses are posted to a completions-only queue
that the completions provider mechanism can consume.

In other words, completion results do not end up in the regular
results queue for other fsac incoming data. This is because
completions must be dealt with synchronously, in contrast to
other fsac commands implemented so far.

Potential pitfalls:

We optimistaclly drain the completions queue when fsac takes time
to respond, but this may potentially cause inconsistent results.