edit build.ps1 to optionally use a config file
- ignore said config file (plugin_dev.json)
- wire up build.ps1 to .sublime-project file

Instead of having to define $STPackagesPath in every session,
it's now possible to store that value in a 'plugin_dev.json' file.
This file lives in .../sublimetext. It's a json object that must
contain a `packages_path` key.

This file is ignored in .gitignore, as it only should be available
locally.

Build.ps1 is now wired up to .sublime-project as the default 'Run'
action. This way, it's possible to press F7 to deploy the files
from within Sublime Text.